backend/requirements.txt
fastapi>=0.115.0
sqlmodel>=0.0.21
uvicorn[standard]>=0.30.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
sqladmin>=0.18.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
itsdangerous>=2.1.0
